requirements/base.txt
argon2-cffi==19.1.0
Babel==2.6.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-appconf==1.0.3
django-autoslug==1.9.4
django-axes==4.5.4
django-classy-tags==0.8.0
django-cleanup==3.2.0
django-cms==3.6.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filer==1.4.4
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-ipware==2.1.0
django-js-asset==1.2.1
django-meta==1.5.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-phonenumber-field==2.2.0
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-social-share==1.3.2
django-treebeard==4.3
djangocms-admin-style==1.3.0
djangocms-attributes-field==1.0.0
djangocms-bootstrap4==1.3.1
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-history==1.0.0
djangocms-icon==1.2.0
djangocms-link==2.3.1
djangocms-modules==0.1.0
djangocms-page-meta==0.8.5
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-transfer==0.1.0
djangocms-video==2.1.1
easy-thumbnails==2.6
entrypoints==0.3
html5lib==1.0.1
idna==2.8
mccabe==0.6.1
oauthlib==3.0.1
phonenumberslite==8.10.6
Pillow==6.0.0
psycopg2==2.7.7
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.24.2
webencodings==0.5.1
